Fixed and moved data retrieval unit tests
