+++ .idea/Finance_Calculator.iml
@@ -5,7 +5,7 @@
       <sourceFolder url="file://$MODULE_DIR$" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/.venv" />
     </content>
-    <orderEntry type="jdk" jdkName="Python 3.13 (finance_calc)" jdkType="Python SDK" />
+    <orderEntry type="jdk" jdkName="Python 3.13 (Finance_Calculator)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
 </module>
